Move TreeBuilder into a shared test helper
Moves the tree fixture builder out of test/browser-specific.js so that other
test files can use it, exporting nextUniqueId to keep a single id counter.
